[wpilibj] Move button tests to button package
Also make CommandTestBase public to facilitate
